requirements.txt
flask
torch
torchvision
gunicorn
uvicorn
pillow
numpy
matplotlib
pandas
scikit-learn
